pyproject.toml
[project]
name = "pdf-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.54.0",
    "pymupdf>=1.26.6",
]
